server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
strawberry-graphql[fastapi]==0.235.0
prometheus-client==0.20.0
pydantic==1.10.13
